Add possibility to directly switch between download plugin to remove plugin
Follow: https://github.com/OpenPLi/enigma2/commit/71a5e7b77095a94f21d715657c189812a9617d39

+++ lib/python/Screens/PluginBrowser.py
@@ -330,10 +330,8 @@ class PluginDownloadBrowser(Screen):
 		self.onChangedEntry = []
 		self["list"].onSelectionChanged.append(self.selectionChanged)
 
-		if self.type in (self.DOWNLOAD, self.MANAGE):
-			self["text"] = Label(_("Downloading plugin information. Please wait..."))
-		elif self.type == self.REMOVE:
-			self["text"] = Label(_("Getting plugin information. Please wait..."))
+		self["text"] = Label(_("Downloading plugin information. Please wait...") if self.type == self.DOWNLOAD else _("Getting plugin information. Please wait..."))
+		self["key_red" if self.type == self.DOWNLOAD else "key_green"] = Label(_("Remove plugins") if self.type == self.DOWNLOAD else _("Download plugins"))
 
 		self["key_red"] = StaticText(_("Close"))
 		self["key_green"] = StaticText("")
@@ -348,10 +346,11 @@ class PluginDownloadBrowser(Screen):
 			"cancel": (self.requestClose, _("Cancel / Close the screen")),
 			"ok": (self.go, _("Perform install/remove of the selected item"))
 		}, prio=0, description=_("Plugin Manager Actions"))
-
-		self.opkg = "/usr/bin/opkg"
-		self.opkg_install = self.opkg + " install --force-overwrite"
-		self.opkg_remove = self.opkg + " remove --autoremove --force-depends"
+		self["PluginDownloadActions"] = ActionMap(["ColorActions"],	{"red": self.delete} if self.type == self.DOWNLOAD else {"green": self.download})
+		if os.path.isfile('/usr/bin/opkg'):
+			self.opkg = "/usr/bin/opkg"
+			self.opkg_install = self.opkg + " install --force-overwrite"
+			self.opkg_remove = self.opkg + " remove --autoremove --force-depends"
 
 		self.opkgObj = OpkgComponent()
 		self.opkgObj.addCallback(self.opkgCallback)
@@ -454,7 +453,13 @@ class PluginDownloadBrowser(Screen):
 				mbox = self.session.openWithCallback(self.runInstall, MessageBox, _("Do you really want to remove the plugin \"%s\"?") % plugin.name, default=False)
 				mbox.setTitle(_("Remove Plugins"))
 
-	def requestClose(self):
+	def delete(self):
+		self.requestClose(1)
+
+	def download(self):
+		self.requestClose(0)
+
+	def requestClose(self, returnValue=None):
 		if self.plugins_changed:
 			plugins.readPluginList(resolveFilename(SCOPE_PLUGINS))
 		if self.reload_settings:
@@ -464,7 +469,7 @@ class PluginDownloadBrowser(Screen):
 		plugins.readPluginList(resolveFilename(SCOPE_PLUGINS))
 		self.container.appClosed.remove(self.runFinished)
 		self.container.dataAvail.remove(self.dataAvail)
-		self.close()
+		self.close(returnValue)
 
 	def resetPostInstall(self):
 		try:
@@ -1188,9 +1193,14 @@ class PluginBrowserNew(Screen):
 		self.session.openWithCallback(self.PluginDownloadBrowserClosed, PluginDownloadBrowser, PluginDownloadBrowser.DOWNLOAD, self.firsttime)
 		self.firsttime = False
 
-	def PluginDownloadBrowserClosed(self):
-		self.updateList()
-		self.checkWarnings()
+	def PluginDownloadBrowserClosed(self, returnValue):
+		if returnValue == None:
+			self.updateList()
+			self.checkWarnings()
+		elif returnValue == 0:
+			self.download()
+		else:
+			self.delete()
 
 	def openExtensionmanager(self):
 		if fileExists(resolveFilename(SCOPE_PLUGINS, "SystemPlugins/SoftwareManager/plugin.py")):
